refactor: delete kernel.py -> move bits to metric.py, quality.py & update test_decomposition to reflect
